For big-endian version of getulong() macro, cast result to u_int32_t. This macro was written expecting a 32-bit unsigned long, and doesn't work properly on 64-bit systems.  This bug caused vn_stat() to return incorrect values for files larger than 2gb on msdosfs filesystems on 64-bit systems.
PR:		106703
Submitted by:	Axel Gonzalez <loox e-shell net>
MFC after:	3 days

Former-commit-id: 3eed5561b47ead027ffd99dc1949838915f06666

+++ sys/fs/msdosfs/bpb.h
@@ -313,7 +313,7 @@ name|getulong
 parameter_list|(
 name|x
 parameter_list|)
-value|(((u_int8_t *)(x))[0] + (((u_int8_t *)(x))[1]<< 8) \ 			 + (((u_int8_t *)(x))[2]<< 16)	\ 			 + (((u_int8_t *)(x))[3]<< 24))
+value|(u_int32_t)(((u_int8_t *)(x))[0] + (((u_int8_t *)(x))[1]<< 8) \ 			 + (((u_int8_t *)(x))[2]<< 16)	\ 			 + (((u_int8_t *)(x))[3]<< 24))
 end_define
 
 begin_define
